LayoutTests/platform/chromium/fast/events/pageshow-pagehide-on-back-uncached-expected.txt

main frame - has 1 onunload handler(s)
ALERT: window.onload
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: Blocked alert('window.onpagehide, target = [object HTMLDocument], persisted = false') during pagehide.
main frame "pageshow/pagehide" - has 1 onunload handler(s)
ALERT: window.onload
ALERT: window.onpageshow, target = [object HTMLDocument], persisted = false
Test pageshow/pagehide event behavior when navigating back to an uncached page.
